+++ truffle/CHANGELOG.md
@@ -14,6 +14,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * GR-51172 Add `CompilerDirectives.ensureAllocatedHere` to mark an allocation as non-movable. This allows language developers to mark special allocations as non-optimizable to allow better control for allocations potentially throwing OutOfMemoryErrors.
 * GR-52930 Added `TruffleLanguage.Env.getHostLanguage()` returning the host language info.
 * GR-52930 Added `TruffleLanguage.Env.getScope(LanguageInfo)` returning the top scope object of the language.
+* GR-28866 Added `TruffleLanguage.Env.getScopePublic(LanguageInfo)` and `TruffleLanguage.Env.getScopeInternal(LanguageInfo)` to allow languages direct access to other language scopes to implement new polyglot builtins.
 
 ## Version 24.0.0